# public/openapi.yaml
openapi: 3.1.0
info:
  title: HeaderGuard API
  version: 2.0.0
  description: |
    **HeaderGuard** is an enterprise-grade HTTP security header scanner & External Attack Surface Management (EASM) platform.

    This API enables you to scan websites for security header compliance, retrieve detailed security audit reports,
    manage continuous monitoring schedules, and administer user accounts — all from a single, powerful REST interface.

    ## Key Capabilities

    | Feature | Description |
    |---|---|
    | **Header Scanning** | Analyze 8+ critical HTTP security headers with severity ratings |
    | **EASM Analysis** | SSL/TLS, DNS, infrastructure, technology stack, WHOIS, and more |
    | **Compliance Checking** | GDPR, PCI-DSS, OWASP, and NIST compliance assessments |
    | **Vulnerability Detection** | Nuclei, Nmap, Nikto integration for deep vulnerability scanning |
    | **Continuous Monitoring** | Daily/weekly automated scans with email alerting on regressions |
    | **Domain Verification** | File-based domain ownership verification before scanning |

    ## Scoring System

    | Score Range | Grade |
    |---|---|
    | 90–100 | A+ |
    | 80–89 | A |
    | 70–79 | B |
    | 60–69 | C |
    | 50–59 | D |
    | < 50 | F |

    Each header is weighted. A correctly configured header earns full weight; a weak header earns 30% weight; a missing header earns 0.

    ## Rate Limiting

    - **Authenticated users**: Daily quota (default 20 requests/day, admin 27/day)
...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Pass a JWT token obtained from `/api/auth/login` in the `Authorization` header.

        ```
        Authorization: Bearer <token>
        ```
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: |
        An HTTP-only `token` cookie is automatically set on login.
        Browser-based clients use this method automatically.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: |
        Pass your API key (prefixed with `hg_`) in the `x-api-key` header.

        ```
        x-api-key: hg_abc123def456...
        ```
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "Authentication required. Please log in to perform scans."
        code:
          type: string
          description: Machine-readable error code
          example: "UNAUTHORIZED"
        success:
          type: boolean
          example: false

    # ── Auth Schemas ─────────────────────────────────────────────────
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
        password:
          type: string
          format: password
          minLength: 6
          example: "securePassword123"
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        token:
          type: string
          description: JWT authentication token
          example: "eyJhbGciOiJIUzI1NiIs..."
        user:
          $ref: '#/components/schemas/UserSummary'
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          example: "667a1b2c3d4e5f6a7b8c9d0e"
        email:
          type: string
          format: email
          example: "user@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
...
paths:
...
  /auth/resend-otp:
    post:
      tags: [Authentication]
      summary: Resend OTP verification code
      description: |
        Generates and sends a new 6-digit OTP verification code to an unverified user's registered email address.
      operationId: resendOtp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
      responses:
        '200':
          description: Verification code resent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  email:
                    type: string
                    example: "user@example.com"
                  message:
                    type: string
                    example: "A new verification code has been sent to your email."
        '400':
          description: Validation error or account already verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: User account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Log in
      description: |
        Authenticates a user with email and password. Returns a JWT token and sets an HTTP-only session cookie.
        Unverified accounts will receive a `403` error.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HTTP-only session cookie
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Account not verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...